+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/util/Versions.kt
@@ -7,8 +7,11 @@ import org.gradle.util.GradleVersion
 internal object AgpVersions {
   val CURRENT: SemVer = SemVer.parse(Version.ANDROID_GRADLE_PLUGIN_VERSION)
   val VERSION_7_4_0: SemVer = SemVer.parse("7.4.0-rc01")
+  val VERSION_8_3_0: SemVer = SemVer.parse("8.3.0")
 
   fun isAGP74(current: SemVer) = current >= VERSION_7_4_0
+
+  fun isAGP83(current: SemVer) = current >= VERSION_8_3_0
 }
 
 internal object GradleVersions {