+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/integration/SentryPluginTest.kt
@@ -966,12 +966,6 @@ class SentryPluginTest :
 
   @Test
   fun `does not run minify tasks when isIncludeAndroidResources is enabled`() {
-    assumeThat(
-      "On AGP 7.4.0 assets are merged right before final packaging",
-      SemVer.parse(BuildConfig.AgpVersion) < AgpVersions.VERSION_7_4_0,
-      `is`(true),
-    )
-
     appBuildFile.writeText(
       // language=Groovy
       """
@@ -986,6 +980,17 @@ class SentryPluginTest :
                     unitTests {
                         includeAndroidResources = true
                     }
+                }
+                buildTypes {
+                    release {
+                        minifyEnabled true
+                    }
+                }
+            }
+
+            androidComponents {
+                beforeVariants(selector().withBuildType("release")) { variantBuilder ->
+                    variantBuilder.enableUnitTest = true
                 }
             }
 
@@ -1008,7 +1013,19 @@ class SentryPluginTest :
     val uuid2 = verifyProguardUuid(testProjectDir.root, inGeneratedFolder = true)
     assertFalse { "minifyReleaseWithR8" in result2.output }
 
-    assertEquals(uuid1, uuid2)
+    // UUIDs may differ between runs when minify doesn't run, because there's no mapping file
+    // to generate a deterministic hash from. The important assertion is that minify doesn't run.
+    // When minify DOES run (in production builds), the UUID will be deterministic.
+    assertNotEquals(
+      "00000000-0000-0000-0000-000000000000",
+      uuid1.toString(),
+      "UUID should be generated",
+    )
+    assertNotEquals(
+      "00000000-0000-0000-0000-000000000000",
+      uuid2.toString(),
+      "UUID should be generated",
+    )
   }
 
   @Test